+++ src/lib.rs
@@ -1,13 +1,11 @@
 mod _daachorse;
 mod _cedarwood;
-mod extract_keywords;
 mod r#match;
 mod text_range;
 mod behavior_for_unmatched;
 mod utils;
 mod error;
 
-pub use extract_keywords::*;
 pub use r#match::*;
 pub use text_range::*;
 pub use behavior_for_unmatched::*;