+++ src/match.rs
@@ -3,19 +3,19 @@ use crate::TextRange;
 #[derive(Debug, Clone, PartialEq)]
 pub struct Match {
     range: TextRange,
-    value: Option<f64>,
+    index_of_patterns: Option<u32>,
 }
 
 impl Match {
-    pub fn new(range: TextRange, value: Option<f64>) -> Self {
-        Self { range, value }
+    pub fn new(range: TextRange, index_of_patterns: Option<u32>) -> Self {
+        Self { range, index_of_patterns }
     }
 
     pub fn range(&self) -> TextRange {
         self.range
     }
 
-    pub fn value(&self) -> Option<f64> {
-        self.value
+    pub fn index_of_patterns(&self) -> Option<u32> {
+        self.index_of_patterns
     }
 }